Implement new JSON-RPC API commands 'getNetTotals(..)' and 'getNetworkHashPs(..)' in 'BtcdClient'; create new domain object 'NetworkTotals'; fix 2 overflow bugs in domain classes 'MemPoolInfo' and 'MiningInfo'; update 'JsonPrimitiveParser'; update several classes in the 'btcd-cli4j-examples' submodule

+++ core/src/main/java/com/neemre/btcdcli4j/core/domain/MemPoolInfo.java
@@ -18,7 +18,7 @@ import com.fasterxml.jackson.annotation.JsonInclude.Include;
 @JsonInclude(Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class MemPoolInfo extends Entity {
-	
+
 	private Integer size;
-	private Integer bytes;
+	private Long bytes;
 }
